(7.2) renames and cleaning
