+++ test/Protractor.config.js
@@ -1,7 +1,7 @@
 exports.config = {
   seleniumAddress: 'http://localhost:4444/wd/hub',/*
 	capabilities: { 'browserName': 'chrome' },*/
-  specs: ['client/e2e/spec.js'],
+  specs: ['client/e2e/e2eSpec.js'],
   jasmineNodeOpts: {
     defaultTimeoutInterval: 15000
   }
