+++ test/client/e2e/e2eSpec.js
@@ -2,17 +2,16 @@
 
 describe('onboarding project', function() {
 
-	var ngMockE2E = require('ng-mock-e2e');
+	//var ngMockE2E = require('ng-mock-e2e');
 	var mocks;
-	var $httpBackend = ngMockE2E.$httpBackend;
+	//var $httpBackend = ngMockE2E.$httpBackend;
 	var url = 'http://localhost:1820/#/usersList/';
 	var endpoint = 'http://localhost:24149/users';
 
 	beforeEach(function() {
-		mocks = require('./mocks');
+		mocks = require('./e2eMocks');
 		browser.addMockModule('userMock', mocks.mockFunction);
 		browser.get(url);
-
 	});
 
 	it('loads the users list page', function() {
